adding gen ai factory adding external login to proxy
